Auction System Fix
The existing auction system is broken.  It references a static location
for the auction file which is no longer updated.  The new method is to
request a JSON object that contains the URL for you to use for the
auction file.  This URL does change and as such must be queried each
time.